+++ courses/tasks.py
@@ -2,21 +2,15 @@
 Tasks for the courses app
 """
 
-import json
 import logging
-import re
-import time
-from datetime import datetime, timedelta
+from datetime import timedelta
 
-import requests
 from django.conf import settings
 from django.db.models import Q
 from requests.exceptions import HTTPError
-from wagtail.models import Page
 
-from cms.models import ExternalCoursePage, CourseIndexPage, LearningOutcomesPage, WhoShouldEnrollPage
-from courses.constants import EMERITUS_REPORT_NAMES
-from courses.models import CourseRun, CourseRunCertificate, Course, Platform, CourseTopic
+from courses.models import CourseRun, CourseRunCertificate
+from courses.sync_external_courses.api import fetch_emeritus_courseruns, update_emeritus_courseruns
 from courses.utils import (
     ensure_course_run_grade,
     process_course_run_grade_certificate,
@@ -117,231 +111,7 @@ def sync_courseruns_data():
 
 
 # @app.task
-def task_sync_emeritus_courses():
-    """Task to sync courses for Emeritus"""
-    data = fetch_emeritus_course_data()
-    update_external_courses(data)
-
-
-def update_external_courses(data):
-    platform = Platform.objects.get(name="Emeritus")
-    for external_course_run in data.get("rows", []):
-        course_title = external_course_run.get("program_name")
-        course_code = external_course_run.get("course_code")
-        course_run_code = external_course_run.get("course_run_code")
-        if not (course_title and course_code and course_run_code):
-            continue
-
-        course = Course.objects.filter(
-            title__iexact=course_title,
-            platform=platform,
-            is_external=True,
-        ).first()
-        course_readable_id = generate_course_readable_id(course_code)
-        if not course:
-            # if exact matching failed, we will do partial matching
-            partial_matching_courses = Course.objects.filter(
-                title__icontains=course_title, platform=platform, is_external=True
-            )
-            partial_matching_course_count = len(partial_matching_courses)
-            if partial_matching_course_count == 1:
-                # A course found in partial matching
-                course = partial_matching_courses[0]
-            elif partial_matching_course_count > 1:
-                # More than 1 course found in partial matching
-                log.error(
-                    "Found duplicate courses when doing a partial match for {}. Skipping...".format(course_title)
-                )
-                continue
-            elif partial_matching_course_count == 0:
-                # No courses found in partial matching
-                log.info("No partial matching course found for title: {}. Creating a new one...".format(course_title))
-
-                course = Course.objects.create(
-                    title=course_title,
-                    readable_id=course_code,
-                    platform=platform,
-                    live=True,
-                    is_external=True,
-                )
-
-        # if course.readable_id != course_code:
-        #     course.readable_id = course_code
-        #     course.save()
-
-        start_date_str = external_course_run.get("start_date", None)
-        end_date_str = external_course_run.get("end_date", None)
-        start_date = datetime.strptime(start_date_str, "%Y-%m-%d") if start_date_str else None
-        end_date = datetime.strptime(end_date_str, "%Y-%m-%d") if end_date_str else None
-
-        course_run_code = external_course_run.get("course_run_code")
-        course_run_tag = generate_external_course_run_tag(course_run_code)
-        course_run_courseware_id = generate_external_course_run_courseware_id(course_run_tag, course_readable_id)
-
-        matching_course_run = course.courseruns.filter(
-            start_date__date=start_date.date(), end_date__date=end_date.date()
-        ).first()
-        if not matching_course_run:
-            CourseRun.objects.create(
-                course=course,
-                title=course_title,
-                courseware_id=course_run_courseware_id,
-                run_tag=course_run_tag,
-                start_date=start_date,
-                end_date=end_date,
-                live=True if not end_date or datetime.now() < end_date else False,
-            )
-        # elif matching_course_run.courseware_id != course_run_code or matching_course_run.run_tag != course_run_tag:
-        #     matching_course_run.courseware_id = course_run_code
-        #     matching_course_run.run_tag = course_run_tag
-        #     matching_course_run.save()
-
-        course_page = getattr(course, "externalcoursepage", None)
-        if not course_page:
-            course_page = ExternalCoursePage(
-                title=course_title,
-                external_marketing_url=external_course_run.get("landing_page_url", ""),
-                subhead="Delivered in collaboration with Emeritus.",
-                duration=f"{external_course_run.get('total_weeks')} Weeks",
-                format=external_course_run.get("format"),
-                description=external_course_run.get("description")
-            )
-            course_index_page = Page.objects.get(id=CourseIndexPage.objects.first().id).specific
-            course_index_page.add_child(instance=course_index_page)
-            course_page.save()
-
-            if external_course_run.get("Category"):
-                topic = CourseTopic.objects.get_or_create(name=external_course_run.get("Category"))
-                course_page.topics.add(topic)
-
-            if external_course_run.get("learning_outcomes"):
-                create_learning_outcomes_page(course_page, external_course_run.get("learning_outcomes"))
-
-            if external_course_run.get("program_for"):
-                create_who_should_enroll_in_page(course_page, external_course_run.get("program_for"))
-
-
-def generate_course_readable_id(external_course_code):
-    if not external_course_code:
-        return ""
-
-    external_course_tag = external_course_code.split("-")[1]
-    return f"course-v1:xPRO+{external_course_tag}x"
-
-
-def generate_external_course_run_tag(course_run_code):
-    return re.search(r"[0-9]{2}-[0-9]{2}#[0-9]+$", course_run_code).group(0)
-
-
-def generate_external_course_run_courseware_id(course_run_tag, course_readable_id):
-    return f"{course_readable_id}+{course_run_tag}"
-
-
-def create_who_should_enroll_in_page(course_page, who_should_enroll_string):
-    who_should_enroll_list = who_should_enroll_string.strip().split("\r\n")
-    who_should_enroll_list = [item.replace("●", "").strip() for item in who_should_enroll_list][1:]
-    who_should_enroll_page = WhoShouldEnrollPage(
-        heading="WHO SHOULD ENROLL",
-        content=json.dumps([{"type": "item", "value": who_should_enroll_item} for who_should_enroll_item in who_should_enroll_list])
-    )
-    course_page.add_child(who_should_enroll_page)
-    who_should_enroll_page.save()
-
-
-def create_learning_outcomes_page(course_page, outcomes_string):
-    learning_outcomes = outcomes_string.strip().split("\r\n")
-    learning_outcomes = [outcome.replace("●", "").strip() for outcome in learning_outcomes][1:]
-    learning_outcome_page = LearningOutcomesPage(
-        heading="WHAT YOU WILL LEARN",
-        sub_heading="MIT xPRO is collaborating with online education provider Emeritus to deliver "
-                    "this online course. By clicking LEARN MORE, you will be taken to a page where "
-                    "you can download the brochure and apply to the program via Emeritus.",
-        outcome_items=json.dumps([{"type": "outcome", "value": outcome} for outcome in learning_outcomes])
-    )
-    course_page.add_child(instance=learning_outcome_page)
-    learning_outcome_page.save()
-
-
-def fetch_emeritus_course_data(n_days=1):
-    """Fetch Emeritus course data"""
-    api_base_url = settings.EMERITUS_API_BASE_URL
-    api_key = settings.EMERITUS_API_KEY
-    log.info(f"Starting api extract for {n_days}...")
-    # Both reports have a last-modified timestamp field.
-    # Use this to limit the results to only relevant records.
-    last_date = datetime.now()
-    first_date = last_date - timedelta(n_days)
-
-    # Get a list of available queries
-    queries = requests.get(f"{api_base_url}/api/queries?api_key={api_key}", timeout=60)
-    queries.raise_for_status()
-
-    for report in queries.json()["results"]:
-        # Check if query is in list of desired reports
-        if report["name"] not in EMERITUS_REPORT_NAMES:
-            # If not, continue.
-            log.info(
-                "Report: {} not specified for extract...skipping".format(report["name"])
-            )
-            continue
-
-        log.info("Requesting data for {}...".format(report["name"]))
-
-        # Make a post request for the query found.
-        # This will return either:
-        #   a) A query_result if one is cached for the parameters set, or
-        #   b) A Job object.
-        data = requests.post(
-            f"{api_base_url}/api/queries/{report['id']}/results?api_key={api_key}",
-            data=json.dumps(
-                {
-                    "parameters": {
-                        "date_range": {"start": f"{first_date}", "end": f"{last_date}"}
-                    }
-                }
-            ),
-        )
-
-        data.raise_for_status()
-        data = data.json()
-
-        if "job" in data.keys():
-            # If a job is returned, we will poll until status = 3 (Success)
-            # Status values 1 and 2 correspond to in-progress,
-            # while 4 and 5 correspond to Failed, and Canceled, respectively.
-            job_id = data["job"]["id"]
-            log.info(f"Job id: {job_id} found...waiting for completion...")
-            while True:
-                # Get the status of the job-id returned by the initial request.
-                job_status = requests.get(
-                    f"{api_base_url}/api/jobs/{job_id}?api_key={api_key}"
-                )
-                job_status.raise_for_status()
-                job_status = job_status.json()
-
-                if job_status["job"]["status"] == 3:
-                    # If true, the query_result is ready to be collected.
-                    log.info("Job complete...requesting results...")
-                    query_resp = requests.get(
-                        f"{api_base_url}/api/query_results/{job_status['job']['query_result_id']}?api_key={api_key}"
-                    )
-                    query_resp.raise_for_status()
-                    data = query_resp.json()
-                    break
-                elif job_status["job"]["status"] in [4, 5]:
-                    # Error
-                    log.error("Job failed!")
-                    break
-                else:
-                    # Continue waiting until complete.
-                    log.info("Job not yet complete... sleeping for 2 seconds...")
-                    time.sleep(2)
-
-        if "query_result" in data.keys():
-            # Check that query_reesult is in the data payload.
-            # Write result as json to the specified directory, using
-            # the report title as the filename.
-            results = dict(data["query_result"]["data"])
-            return results
-        else:
-            log.error("Something unexpected happened!")
+def task_sync_emeritus_courseruns():
+    """Task to sync Emeritus courseruns"""
+    emeritus_courseruns = fetch_emeritus_courseruns()
+    update_emeritus_courseruns(emeritus_courseruns)
